% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umlaute.R
\name{replace_umlaute}
\alias{replace_umlaute}
\title{Replace German Umlaute in character vector or RStudio document}
\usage{
replace_umlaute(x = rstudioapi::getActiveDocumentContext())
}
\arguments{
\item{x}{A character vector or rstudioapi document context.}
}
\description{
Replace German Umlaute in character vector or RStudio document
}
